refactor(cli): extract mod tests to sibling file
